% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datiRendistat-package.R
\docType{package}
\name{datiRendistat-package}
\alias{datiRendistat}
\alias{datiRendistat-package}
\title{datiRendistat: I dati dell'indagine Rendistat}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
