% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datiRendistat-package.R
\docType{package}
\name{datiRendistat-package}
\alias{datiRendistat}
\alias{datiRendistat-package}
\title{datiRendistat: I dati dell'indagine Rendistat}
\description{
I dati dell'indagine RendiSTAT condotta nel 2016 Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
